[+] zephyr: command 'pnp' was added. [+] msvc 64-bits target added and proper fixes for implicit conversion size_t to int. [*] header file with license related to risc-v were modified.

+++ zephyr/lib/libc/minimal/include/bits/size_t.h
@@ -31,7 +31,7 @@ typedef unsigned int size_t;
 #elif defined(__NIOS2__)
 typedef unsigned int size_t;
 #elif defined(CONFIG_RISCV64)
-typedef unsigned int  size_t;
+typedef unsigned long long  size_t;
 #else
 #error "The minimal libc library does not recognize the architecture!\n"
 #endif
